// src/app/models/eventlog.model.ts
export class Eventlog {
    eventLogId: string;
    description: string;
    active: true;
    tresholdMetTime: string;
    organizationId: string;
    locationId: string;
    assetId: string;
    organizationName: string;
    locationName: string;
    assetName: string;
    signalValue: string;
    alarmStatus: string;
    alertName: string;
    singalName: string;
    sensorName: string;

}